gh-2772 Re-modeling ecdsa signatures
